content: Use ref instead of relref for internal links
Hugo 0.45 changed the behavior of ref and relref. The blog post, docs,
and error messages are all fairly confusing to be honest, but changing
the paths to simply include the document name seems to work, though it
appears there is a non-standard case when resolving a reference from a
post in another language. In that case the only ref syntax that worked
was the longer one that uses "path" and "lang".

Switching from relref to ref is probably unnecessary, but I wanted the
links to be absolute and never really noticed that they weren't.

See: https://gohugo.io/news/0.45-relnotes/