config.toml: Add Farsi translation of About block
This is really tricky to work with in text editors because of the
mixing of Latin and Persio–Arabic scripts, not to mention Markdown!
Note that we need to have the two spaces there at the end of the
line so that we get a hard line break in markdown.